requirements.txt
Django==1.11.8
djangorestframework==3.7.3
jsonfield==2.0.2
pkg-resources==0.0.0
pytz==2017.3
